Add files via upload
This version of the code includes changes to the drive code.  The only change we made was to cube the joystick inputs making it less sensitive at a lower input.  The drivers viewed this change as neutral.  It changed their driving habits but wasn't overwhelmingly positive or negative.  It would just take some time to get used to.