Add files via upload
The changes to this code include: changing the encoder count because the gear ratio from the output shaft to the encoder had been changed and putting in older paths in autonomous to see if they worked with the new encoder count.  The old paths closely resembled the path a human would walk, but we found that the robot wasn't reaching its target so we began to overcompensate.  The changes were made to this version in preparation for the boiler bot competition.